# src/main/resources/static/swagger.yml
swagger: '2.0'
info:
  description: ACCREDITATION VUE API Document.
  version: '1.0'
  title: ACCREDITATION VUE API
  contact:
    name: evildoer
    url: 'https://baidu.com'
    email: 123456@qq.com
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
host: '39.108.220.199:8888'
basePath: /
...
paths:
...
  /v1/course/deleteByIds:
    delete:
      tags:
        - course-controller
      summary: 通过id列表批量删除课程
      operationId: deleteByIdsUsingDELETE
      produces:
        - '*/*'
      parameters:
        - in: body
          name: ids
          description: ids
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ServerResponse'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  '/v1/course/{id}':
    get:
      tags:
        - course-controller
      summary: 通过id查找课程
      operationId: queryUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ServerResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    put:
      tags:
        - course-controller
      summary: 通过id更新课程
      operationId: updateByIdUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: course
          description: course
          required: true
          schema:
            $ref: '#/definitions/Course'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ServerResponse'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    delete:
      tags:
        - course-controller
      summary: 通过id删除课程
      operationId: deleteByIdUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ServerResponse'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
...
definitions:
  Course:
    type: object
    properties:
      code:
        type: string
        description: 课程代码
      createTime:
        type: string
        format: date-time
        description: 创建时间
      credit:
        type: number
        format: double
        description: 课程对应学分
      id:
        type: integer
        format: int32
        description: id
      name:
        type: string
        description: 课程名称
      nature:
        type: string
        description: 课程性质
      updateTime:
        type: string
        format: date-time
        description: 更新时间
    title: Course
...
  ServerResponse:
    type: object
    properties:
      data:
        type: object
      msg:
        type: string
      status:
        type: integer
        format: int32
      success:
        type: boolean
    title: ServerResponse
...